# public/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CribsTalk API",
    "description": "Complete API documentation for the CribsTalk platform — a social network with marketplace, messaging, groups, stories, and more.",
    "version": "1.0.0",
    "contact": {
      "name": "CribsTalk Team"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "Default API server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Supabase JWT token from authenticated session"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        }
      },
...
    },
    "responses": {
...
      "ServerError": {
        "description": "Internal server error",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      }
    }
  },
  "paths": {
...
    "/videosdk/token": {
      "get": {
        "tags": ["VideoSDK"],
        "summary": "Get VideoSDK JWT token",
        "parameters": [
          { "name": "roomId", "in": "query", "schema": { "type": "string" } },
          { "name": "userId", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "JWT token", "content": { "application/json": { "schema": { "type": "object", "properties": { "token": { "type": "string" } } } } } },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      },
      "post": {
        "tags": ["VideoSDK"],
        "summary": "Generate VideoSDK JWT token",
        "requestBody": {
          "content": { "application/json": { "schema": { "type": "object", "properties": { "roomId": { "type": "string" }, "userId": { "type": "string" } } } } }
        },
        "responses": {
          "200": { "description": "JWT token", "content": { "application/json": { "schema": { "type": "object", "properties": { "token": { "type": "string" } } } } } },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
...
  }
}